this.isModified적용, Owner인 사람만 비디오 관련 작업이 가능하도록 수정

+++ src/controllers/videoController.js
@@ -21,19 +21,27 @@ export const watch = async(req, res) => {
 
 export const getEdit = async(req, res) => {
     const {id} = req.params; // ES6
+    const {user: {_id}} = req.session;
     const video = await Video.findById(id);
     if(video === null){
         return res.status(404).render("404", {pageTitle:"Video not found."});
+    }
+    if(String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
     }
     return res.render("edit", {pageTitle:`Edit: ${video.title}`, video});
 }
 
 export const postEdit = async(req, res) => {
+    const {user: {_id}} = req.session;
     const {id} = req.params; // ES6
     const {title, description, hashtags} = req.body;
     const video = await Video.exists({_id:id});
     if(video === null){
         return res.render("404", {pageTitle:"Video not found."});
+    }
+    if(String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title,
@@ -53,13 +61,16 @@ export const postUpload = async (req, res) => {
     const {path} = req.file;
     const {title, description, hashtags} = req.body;
     try{
-        await Video.create({
+        const newVideo = await Video.create({
             title, 
             description,
             fileUrl:path,
             owner:_id,
             hashtags: Video.formatHashtags(hashtags),
         });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
         return res.redirect("/");
     }
     catch(error){
@@ -69,8 +80,22 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
     const {id} = req.params;
-    // delete video
+    const {user: {_id}} = req.session;
+    const video = await Video.findById(id);
+    const user = await User.findById(_id);
+    if(!video)
+    {
+        res.render("404", {pageTitle:"Video not found."});
+    }
+    
+    if(String(video.owner) !== String(_id)) {
+        console.log("예외처리");
+        return res.status(403).redirect("/");
+    }
+
     await Video.findByIdAndDelete(id);
+    user.videos.splice(user.videos.indexOf(id), 1);
+    user.save();
     return res.redirect("/");
 }
